test: add unit coverage for fabric services state v3
